% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-matrix-rowSums.r
\name{colSums}
\alias{colSums}
\alias{rowSums.dual}
\alias{rowSums}
\alias{colMeans}
\alias{rowMeans}
\alias{rowSums,dual-method}
\alias{colSums.dual}
\alias{colSums,dual-method}
\alias{rowMeans.dual}
\alias{rowMeans,dual-method}
\alias{colMeans.dual}
\alias{colMeans,dual-method}
\title{Row and column sums and means}
\usage{
rowSums.dual(x, na.rm = FALSE, dims = 1, ...)

\S4method{rowSums}{dual}(x, na.rm = FALSE, dims = 1, ...)

colSums.dual(x, na.rm = FALSE, dims = 1, ...)

\S4method{colSums}{dual}(x, na.rm = FALSE, dims = 1, ...)

rowMeans.dual(x, na.rm = FALSE, dims = 1, ...)

\S4method{rowMeans}{dual}(x, na.rm = FALSE, dims = 1, ...)

colMeans.dual(x, na.rm = FALSE, dims = 1, ...)

\S4method{colMeans}{dual}(x, na.rm = FALSE, dims = 1, ...)
}
\arguments{
\item{x}{a dual matrix or array}

\item{na.rm}{if 'TRUE', missing values are removed}

\item{dims}{which dimensions are regarded as rows and cols}

\item{...}{extra parameters (ignored)}
}
\value{
a dual object (usually a dual vector).
}
\description{
Method extending to dual matrices the corresponding methods for dual matrices.
}
\examples{
x <- dual( c(1,2) )
x <- cbind(x, 2*x+1)
rowSums(x)
d(rowSums(x), "x1")

}
