% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-compare.r
\name{Comparison}
\alias{Comparison}
\alias{Compare,dual,ANY-method}
\title{Comparison Operators}
\usage{
\S4method{Compare}{dual,ANY}(e1, e2)
}
\arguments{
\item{e1}{object of class 'dual' or 'numeric'}

\item{e2}{object of class 'dual' or 'numeric'}
}
\value{
a logical vector
}
\description{
Comparison operators for objects of class 'dual'
}
\details{
usual comparison operators, ignoring derivatives valuesa
}
