% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slides.R
\name{create_slides}
\alias{create_slides}
\title{Create slides from .qmd files}
\usage{
create_slides(...)
}
\arguments{
\item{...}{a .qmd-file}
}
\value{
both a .html and a .pdf version
}
\description{
Create slides from .qmd files
}
