% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-replace.r, R/dual-subsetting.r
\name{Extract}
\alias{Extract}
\alias{[<-,dual,index,index,dual-method}
\alias{[<-,dual,missing,index,dual-method}
\alias{[<-,dual,index,missing,dual-method}
\alias{[<-,dual,missing,missing,dual-method}
\alias{[<-,dual,index,index,logicalOrNumericOrArray-method}
\alias{[<-,dual,missing,index,logicalOrNumericOrArray-method}
\alias{[<-,dual,index,missing,logicalOrNumericOrArray-method}
\alias{[<-,dual,missing,missing,logicalOrNumericOrArray-method}
\alias{[,dual,index,index-method}
\alias{[,dual,missing,index-method}
\alias{[,dual,index,missing-method}
\alias{[,dual,missing,missing-method}
\title{Extract or replace parts of an object}
\usage{
\S4method{[}{dual,index,index,dual}(x, i, j, ...) <- value

\S4method{[}{dual,missing,index,dual}(x, i, j, ...) <- value

\S4method{[}{dual,index,missing,dual}(x, i, j, ...) <- value

\S4method{[}{dual,missing,missing,dual}(x, i, j, ...) <- value

\S4method{[}{dual,index,index,logicalOrNumericOrArray}(x, i, j, ...) <- value

\S4method{[}{dual,missing,index,logicalOrNumericOrArray}(x, i, j, ...) <- value

\S4method{[}{dual,index,missing,logicalOrNumericOrArray}(x, i, j, ...) <- value

\S4method{[}{dual,missing,missing,logicalOrNumericOrArray}(x, i, j, ...) <- value

\S4method{[}{dual,index,index}(x, i, j, ..., drop = TRUE)

\S4method{[}{dual,missing,index}(x, i, j, ..., drop = TRUE)

\S4method{[}{dual,index,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{dual,missing,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{dual object}

\item{i, j}{indices of elements to extract or replace}

\item{...}{supplementary indices (for arrays)}

\item{value}{replacement value}

\item{drop}{for dual matrices or array.}
}
\value{
returns a dual object (the semantic is the same as base extraction and replacement methods).
}
\description{
Methods for extraction or replacements of parts of dual objects.
}
\examples{
x <- c(1, 2, 3)
x[2] <- dual(4)
x
d(x)
}
