% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_slides.R
\name{hide_answer}
\alias{hide_answer}
\title{Hide answer slides in quarto revealjs-presentations}
\usage{
hide_answer(x, ...)
}
\arguments{
\item{x}{An .Rmd file}

\item{...}{additional arguments passed to \code{quarto::quarto_render()}}
}
\value{
a HTML file
}
\description{
Hide answer slides in quarto revealjs-presentations
}
