% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-rep-t-aperm.r
\name{rep}
\alias{rep}
\alias{rep.dual}
\title{Replicate elements of a dual vector}
\usage{
\method{rep}{dual}(x, ...)
}
\arguments{
\item{x}{a dual vector}

\item{...}{extra parameters (typically, 'times', 'length.out' or 'each')}
}
\value{
A dual object.
}
\description{
A method extending `rep` to dual objects
}
\examples{
x <- rep( dual(1:2), each = 4 )
x
d(x)

}
