% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-concat-binding.r
\name{c}
\alias{c}
\alias{concat0}
\alias{c,numericOrArray-method}
\alias{c-dual}
\title{Concatenation methods}
\usage{
\S4method{c}{numericOrArray}(x, ...)
}
\arguments{
\item{x}{first object to concatenate}

\item{...}{other objects}
}
\value{
an object of class dual.
}
\description{
Methods have been defined in order to allow the concatenation 
of `dual` objects together and with constant objects.
}
\examples{
x <- dual( 1 )
# concatenation with a constant
x <- c(x, 2)
x
d(x)
# concatenation of dual objects
x1 <- sum(x)
x2 <- sum(x**2)
y <- c(a = x1, b = x2)  # you can use named arguments
y
d(y)
}
