% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-d.r, R/dual-value.r
\name{d}
\alias{d}
\alias{d.dual}
\alias{d.numeric}
\alias{value}
\alias{value.dual}
\alias{value.numeric}
\title{d}
\usage{
d(x, varnames)

value(x)

\method{value}{dual}(x)

\method{value}{numeric}(x)
}
\arguments{
\item{x}{a dual (or numeric) oject}

\item{varnames}{(optional) a vector or varnames to take derivatives along}
}
\description{
Get value or differential of an object.
}
\details{
If `varnames` is provided, either a single derivative or a list of derivatives will be sent back
}
\examples{
x <- dual(c(3,2))
x**2
value(x**2)
d(x**2)
d(x**2, "x1")
# you may use a constant
value(1)
d(1, "x1")

}
