% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_slides.R
\name{show_answer}
\alias{show_answer}
\title{Show answer slides in quarto revealjs-presentations}
\usage{
show_answer(x, ...)
}
\arguments{
\item{x}{An .Rmd file}

\item{...}{}
}
\value{
a HTML file
}
\description{
Show answer slides in quarto revealjs-presentations
}
